+++ src/main/java/org/jugni/apps/pico/data/dao/BorrarRegistroDao.java
@@ -1,6 +1,5 @@
-package org.jugni.apps.pico.dao;
+package org.jugni.apps.pico.data.dao;
 
-import java.util.List;
 
 /**
  * 
@@ -8,6 +7,6 @@ import java.util.List;
  * @version :1.0 
  * @param <T> En la clase que lo implemente se sustituye por la clase entidad
  */
-public interface CrearRegistrosDao<T> {
-     void insertarRegistros(List<T> lista);
+public interface BorrarRegistroDao<T> {
+     void borrarRegistros(T registro);
 }
