+++ src/main/java/org/jugni/apps/pico/data/dao/CrearRegistrosDao.java
@@ -1,5 +1,6 @@
-package org.jugni.apps.pico.dao;
+package org.jugni.apps.pico.data.dao;
 
+import java.util.List;
 
 /**
  * 
@@ -7,6 +8,6 @@ package org.jugni.apps.pico.dao;
  * @version :1.0 
  * @param <T> En la clase que lo implemente se sustituye por la clase entidad
  */
-public interface BorrarRegistroDao<T> {
-     void borrarRegistros(T registro);
+public interface CrearRegistrosDao<T> {
+     void insertarRegistros(List<T> lista);
 }
