+++ src/main/java/org/jugni/apps/pico/data/dao/CuentaTipoDao.java
@@ -1,4 +1,4 @@
-package org.jugni.apps.pico.dao;
+package org.jugni.apps.pico.data.dao;
 
 import org.hibernate.SessionFactory;
 import org.jugni.apps.pico.modelos.CuentaTipo;
@@ -8,5 +8,5 @@ public class CuentaTipoDao extends BaseDao<CuentaTipo, Short> {
   public CuentaTipoDao(SessionFactory factory) {
     super(factory, CuentaTipo.class);
   }
-  
+
 }
