+++ src/main/java/org/jugni/apps/pico/data/dao/ObtenerRegistroDao.java
@@ -1,4 +1,4 @@
-package org.jugni.apps.pico.dao;
+package org.jugni.apps.pico.data.dao;
 
 /**
   * 
@@ -8,6 +8,6 @@ package org.jugni.apps.pico.dao;
  *
  *   Interface ObtenerUnRegistroDao : Interfas para obtener un registro de la base de datos condicionado por la clase registro
  */
-public interface ObtenerRegistroSiDao <T> {
-      T obtenerRegistroSi(T registro);
+public interface ObtenerRegistroDao <T> {
+      T obtenerRegistro();
 }
