+++ src/main/java/org/jugni/apps/pico/vista/swing/CuentaForm.java
@@ -1,6 +1,6 @@
 package org.jugni.apps.pico.vista.swing;
 
-import org.jugni.apps.pico.dao.CuentaTipoImpl;
+import org.jugni.apps.pico.data.dao.CuentaTipoImpl;
 import org.jugni.apps.pico.modelos.CuentaTipo;
 import org.jugni.apps.pico.vista.utils.ButtonContructor;
 import org.jugni.apps.pico.vista.utils.CerrarButton;
