+++ src/main/java/org/jugni/apps/pico/vista/swing/CuentaTipoForm.java
@@ -8,7 +8,7 @@ import java.awt.event.KeyEvent;
 import javax.swing.*;
 import javax.swing.border.LineBorder;
 import javax.swing.border.TitledBorder;
-import org.jugni.apps.pico.dao.CuentaTipoImpl;
+import org.jugni.apps.pico.data.dao.CuentaTipoImpl;
 import org.jugni.apps.pico.modelos.CuentaTipo;
 import org.jugni.apps.pico.vista.utils.ButtonContructor;
 import org.jugni.apps.pico.vista.utils.CerrarButton;
